src/requirements.txt
pandas
numpy
redis
requests
sentence-transformers
python-dotenv
tabulate
ipython